Move order queries business logic to core-orders package #275
