+++ packages/api/resolvers/queries/orders.js
@@ -1,5 +1,5 @@
 import { log } from 'meteor/unchained:core-logger';
-import { Orders, OrderStatus } from 'meteor/unchained:core-orders';
+import { Orders } from 'meteor/unchained:core-orders';
 
 export default function orders(
   root,
@@ -7,17 +7,5 @@ export default function orders(
   { userId }
 ) {
   log(`query orders: ${limit} ${offset} ${includeCarts}`, { userId });
-
-  const selector = {};
-  if (!includeCarts) {
-    selector.status = { $ne: OrderStatus.OPEN };
-  }
-  const options = {
-    skip: offset,
-    limit,
-    sort: {
-      created: -1,
-    },
-  };
-  return Orders.find(selector, options).fetch();
+  return Orders.findOrders({ limit, offset, includeCarts });
 }
